refactor(storage): organize routers into ceph/ and swift/ subdirectories
Move storage routers into separate subdirectories to improve organization and avoid confusion between Ceph and Swift implementations.

Changes:

- Create routers/ceph/ and routers/swift/ subdirectories

- Move containerRouter, objectRouter, ec2CredentialRouter to ceph/

- Move swiftRouter to swift/

- Add index.ts exports for each subdirectory

- Update all import paths in routers and tests

- Update main routers/index.ts to import from subdirectories

All 4397 tests passing after restructure.
