+++ apps/aurora-portal/src/server/Storage/routers/ceph/containerRouter.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from "vitest"
 import { TRPCError } from "@trpc/server"
-import { AuroraPortalContext } from "../../context"
+import { AuroraPortalContext } from "../../../context"
 import { containerRouter } from "./containerRouter"
-import { createCallerFactory, auroraRouter } from "../../trpc"
+import { createCallerFactory, auroraRouter } from "../../../trpc"
 
 // ============================================================================
 // MOCK AWS SDK S3 CLIENT
@@ -10,7 +10,7 @@ import { createCallerFactory, auroraRouter } from "../../trpc"
 
 const mockSend = vi.fn()
 
-vi.mock("../clients/s3Client", () => ({
+vi.mock("../../clients/s3Client", () => ({
   createS3Client: vi.fn(() => ({ send: mockSend })),
 }))
