+++ apps/aurora-portal/src/server/Storage/routers/ceph/index.ts
@@ -1,0 +1,3 @@
+export { containerRouter } from "./containerRouter"
+export { objectRouter } from "./objectRouter"
+export { ec2CredentialRouter } from "./ec2CredentialRouter"
