+++ apps/aurora-portal/src/server/Storage/routers/index.ts
@@ -1,7 +1,5 @@
-import { swiftRouter } from "./swiftRouter"
-import { ec2CredentialRouter } from "./ec2CredentialRouter"
-import { containerRouter } from "./containerRouter"
-import { objectRouter } from "./objectRouter"
+import { swiftRouter } from "./swift"
+import { ec2CredentialRouter, containerRouter, objectRouter } from "./ceph"
 import { auroraRouter } from "../../trpc"
 
 export const objectStorageRouters = {
