+++ apps/aurora-portal/src/server/Storage/routers/swift/index.ts
@@ -1,0 +1,1 @@
+export { swiftRouter } from "./swiftRouter"
